Implemented auto tests for TC_2_03_5
Implemented auto tests for TC_2_03_5
