+++ CaesarTests/CaesarTests.csproj
@@ -83,7 +83,7 @@
     <Compile Include="TC_2_03_3.cs" />
     <Compile Include="TC_2_03_2.cs" />
     <Compile Include="TC_3_06_01.cs" />
-    <Compile Include="TC_2_03_6.cs" />
+    <Compile Include="TC_2_03_5.cs" />
     <Compile Include="TC_2_03_1.cs" />
     <Compile Include="TC_2_01_3.cs" />
     <Compile Include="TC_1_02_02_additional.cs" />
